Update tentative schedule
Basically just removes all the speculative detail from the schedule.
I had added all of that because I thought that was the level of detail
required, but just before I pushed my section, I double-checked the
Marked Project Plan and it said to just have major milestones in the
schedule so I deleted all the detail.